requirements.txt
streamlit==1.36.0
absl-py
aiohappyeyeballs
aiohttp
aiosignal
altair
astunparse
async-timeout
attrs
blinker
Brotli
cachetools
certifi
cffi
charset-normalizer
click
colorama
cryptography
flatbuffers
frozenlist
gast
gitdb
GitPython
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
importlib-metadata
importlib_resources
Jinja2
jsonschema
jsonschema-specifications
keras
Keras-Preprocessing
Markdown
markdown-it-py
MarkupSafe
mdurl
mkl-fft
mkl-random
mkl-service
multidict
narwhals
numpy
oauthlib
opencv-contrib-python
opencv-python
opt-einsum
packaging
pandas
pillow
pip
pkgutil_resolve_name
platformdirs
pooch
protobuf
pyarrow
pyasn1
pyasn1-modules
pycparser
pydeck
Pygments
PyJWT
pyOpenSSL
PySocks
python-dateutil
pytz
referencing
Requests==2.32.3
requests-oauthlib
rich
rpds-py
rsa
scipy
setuptools
six
smmap
streamlit-lottie
tenacity
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
termcolor
toml
tornado
typing_extensions
tzdata
urllib3
watchdog
Werkzeug
wheel
win-inet-pton
wrapt
yarl
zipp